Refactored schema validation to use `JSONSchemedStrategy` for consistency across dict and array-like subjects. Added reusable `to_match_schema` method. Updated documentation and incremented version to 0.7.04.
