+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "hyperiontf"
-version = "0.7.03"
+version = "0.7.04"
 description = "Hyperion Testing Framework"
 authors = ["Oleksiy Bondar <bondaroleksiyandriyovich@gmail.com>"]
 license = "Apache-2.0"
